restructure of CreateRecipe & renaming
